% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkTools.R
\docType{package}
\name{LinkTools-package}
\alias{LinkTools-package}
\alias{LinkTools}
\title{R-package 'LinkTools'}
\description{
Tools for linking data sets.
}
\details{
Once finalized, four steps should be integrated into this package:

a) the preparation of data sets which should be linked, i.e. the
transformation into a comparable format and the assignment of new
information, in particular shared unique identifiers,

b) the merge of data sets based on these identifiers,

c) the encoding or enrichment of the data with different output formats
(data.table, XML or CWB).

d) the package includes a wrapper for the Named Entity Linking of textual
data based on DBpedia Spotlight.
}
\keyword{package}
