% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkTools.R
\docType{data}
\name{stammdaten_wikidatafied_2022_02_01_min}
\alias{stammdaten_wikidatafied_2022_02_01_min}
\title{Stammdaten with WikiData-IDs}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1397 rows and 5 columns.
}
\source{
https://www.bundestag.de/services/opendata (Creation Date 2021-11-04)

https://de.wikipedia.org/wiki/Liste_der_Mitglieder_des_Deutschen_Bundestages_(13._Wahlperiode) (Information Retrieved on 2021-11-23)

https://de.wikipedia.org/wiki/Liste_der_Mitglieder_des_Deutschen_Bundestages_(14._Wahlperiode) (Information Retrieved on 2021-11-23)
}
\usage{
stammdaten_wikidatafied_2022_02_01_min
}
\description{
A minimized version of the Stammdaten of the German Bundestag of the 13th and
14th legislative period with added WikiData IDs retrieved via the Wikidata Query 
Service and added party affiliations specific for the legislative period retrieved 
from Wikipedia. For preparation see bt_stammdaten.
}
\keyword{datasets}
