% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkTools.R
\docType{package}
\name{LinkTools-package}
\alias{LinkTools-package}
\alias{LinkTools}
\title{R-package 'LinkTools'}
\description{
Tools for linking data sets via shared unique identifiers.
}
\details{
Four steps are integrated into this package:

a) the preparation of data sets which should be linked, i.e. the
transformation into a comparable format and the assignment of shared unique
identifiers,

b) the merge of data sets based on these identifiers,

c) the encoding or enrichment of the data with three output formats
(data.table, XML or CWB).

d) the package includes a wrapper for the Named Entity Linking of textual
data based on DBpedia Spotlight.
}
\keyword{package}
